+++ packages/fields/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone/fields
+
 # @keystone-alpha/fields
 
 ## 15.0.0